requirements.txt
Pillow
numpy
psutil
requests
ultralytics
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.60.0
fsspec==2025.9.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==2.13.1
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numpy==1.24.3
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
playsound==1.3.0
polars==1.33.1
progressbar2==4.5.0
protobuf==3.20.3
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pygam==0.10.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-utils==3.9.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf-slim==1.1.0
torch==2.8.0
torchvision==0.23.0
ttkthemes==3.2.2
typing_extensions==4.5.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3

